Moved marketplace initialization into its own helper function.
